Using new ObjectLoader
There are now MeshGroups that contain a bunch of Meshes and Materials.

Not yet finished. There is a bug where we need to pass a path relative to the resources directory instead of an absolute one to load the textures.

Everything only makes use of the first mesh in the MeshGroup right now. So it shouldn't change the structure much.

I modified the ObjectLoader to put in absolute paths for the textures, instead of relative to the MTL. So we should include that code directly in our project, with the necessary liscense information.