3rd person view gun moves with arm
Heavily modified the gnome model. It's head is smaller, legs closer together, left arm further down and separated, right arm forward and separated, and it includes the jetpack.

All guns are now a reasonable size and are all positioned directly in the hand of the gnome. The arm rotates instead of the weapon container. Player now derives from SceneNode instead of BasicMeshGroupNode, but contains a BasicMeshGroupNode called playerModel.