Got two trees looking good.
Added per-node back culling and ambient factors, as well as MTL alpha + alpha map support. The ambient factor is more of an override for the leaves, so it does not come from the MTL file.

The shader now culls any fragment with less than 0.5 alpha. Ideally this would be something like 0.05, but we still aren't sorting our transluscent geometry (painter's algorithm) so it doesn't work well. Also 0.95 was a bit too aggressive because of interpolation and such.

I'm working on a function that randomly generates a tree. We will distribute these around the perimiter once it's done.