Lots of trees lookin' good and fast!
I created a distribution function to place random trees.

I changed the models so that we're not making thousands of draw calls for the leaves.

We no longer do blending on the leaves, since we're discarding fragments with no alpha (and blending is now a manual setting per mesh, not automatically determined from the material).

Changed the ground mesh + UVs and the color of the ground texture, to make it look better with the skybox. The ground is also waaay bigger, but the treeline distracts you and covers the distant horizon pretty well.